proagua_backend/urls.py
# proagua_backend/urls.py
from django.contrib import admin
from django.urls import path, include
from django.conf import settings
from django.conf.urls.static import static
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include('api.urls')),           # ← Tout API nou yo
    # path('api-auth/', include('rest_framework.urls')),  # pou login nan browsable API
    # path('api/token/', TokenObtainPairView.as_view(), name='tokent_obtain_pair'),
    # path('api/token/refresh/', TokenRefreshVieew.as_view(), name='token_refresh'),
]

urlpatterns += [
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
]

if settings.DEBUG:
    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
